added item type to items

+++ src/main/java/com/example/shop/dto/MainItemDto.java
@@ -4,6 +4,8 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter @Setter
 public class MainItemDto {
 
@@ -15,10 +17,10 @@ public class MainItemDto {
 
     private String imgUrl;
 
-    private Integer price;
+    private BigDecimal price;
 
     @QueryProjection
-    public MainItemDto(Long id, String itemName, String itemDetail, String imgUrl, Integer price) {
+    public MainItemDto(Long id, String itemName, String itemDetail, String imgUrl, BigDecimal price) {
         this.id = id;
         this.itemName = itemName;
         this.itemDetail = itemDetail;
