+++ src/test/java/com/example/shop/service/ItemImgServiceTest.java
@@ -17,6 +17,7 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,7 +45,7 @@ class ItemImgServiceTest {
         itemFormDto.setItemName("Test Item");
         itemFormDto.setItemSellStatus(ItemSellStatus.SELL);
         itemFormDto.setItemDetail("This is for testing");
-        itemFormDto.setPrice(1000);
+        itemFormDto.setPrice(BigDecimal.valueOf(1000.0));
         itemFormDto.setStockNumber(100);
 
         List<MultipartFile> multipartFileList = createMultipartFiles();
